Revert "chore: Support modular import"
