+++ .npmignore
@@ -5,5 +5,4 @@ node_modules
 yarn.lock
 jest.config.js
 babel.config.js
-babel.plugin.import.js
 gulpfile.js
