+++ package.json
@@ -33,16 +33,13 @@
     "@pitrix/lego-locale": "^0.1.10",
     "async-validator": "^3.4.0",
     "bulma": "^0.9.0",
+    "classnames": "^2.2.6",
     "flatpickr": "^4.6.6",
     "fuse.js": "^6.4.1",
-    "popper.js": "^1.16.1"
-  },
-  "peerDependencies": {
-    "classnames": "^2.2.6",
     "lodash": "^4.17.20",
-    "prop-types": "^15.7.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "popper.js": "^1.16.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
     "@babel/core": "^7.11.5",
@@ -73,9 +70,9 @@
     "node-sass": "^4.14.1",
     "react-is": "^16.13.1",
     "sass-loader": "^9.0.3",
-    "semantic-release": "^17.2.1",
     "storybook-dark-mode": "^1.0.2",
-    "style-loader": "^1.1.3"
+    "style-loader": "^1.1.3",
+    "semantic-release": "^17.2.1"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/",
