--- src/components/Layout/Column.jsx
@@ -1,6 +1,0 @@
-import Basic from "./Basic";
-import Generator from "../../utils/Generator";
-
-const Column = Generator({ componentCls: "column" })(Basic);
-
-export default Column;
