--- src/components/Layout/Columns.jsx
@@ -1,6 +1,0 @@
-import Basic from "./Basic";
-import Generator from "../../utils/Generator";
-
-const Columns = Generator({ componentCls: "columns" })(Basic);
-
-export default Columns;
