--- src/components/Layout/Level.jsx
@@ -1,6 +1,0 @@
-import Basic from "./Basic";
-import Generator from "../../utils/Generator";
-
-const Level = Generator({ componentCls: "level" })(Basic);
-
-export default Level;
