+++ src/components/Select/Select.jsx
@@ -7,10 +7,8 @@ import Tag from "../Tag";
 import Icon from "../Icon";
 import Loading from "../Loading";
 import Option from "./Option";
-import LocaleProvider from "../LocaleProvider";
 import { getScrollParents } from "../../utils";
-
-const { locale } = LocaleProvider;
+import { locale } from "../LocaleProvider";
 
 export default class Select extends React.Component {
   static propTypes = {
