+++ src/index.js
@@ -8,7 +8,6 @@ export { default as Dropdown } from "./components/Dropdown";
 export { default as Form } from "./components/Form";
 export { default as Icon } from "./components/Icon";
 export { default as Loading } from "./components/Loading";
-export { default as LocaleProvider } from "./components/LocaleProvider";
 export { default as Menu } from "./components/Menu";
 export { default as Notify } from "./components/Notify";
 export { default as Pagination } from "./components/Pagination";
@@ -25,4 +24,5 @@ export { default as Tooltip } from "./components/Tooltip";
 export * from "./components/Checkbox";
 export * from "./components/Input";
 export * from "./components/Layout";
+export * from "./components/LocaleProvider";
 export * from "./components/Radio";
